Add new TimeHandler class
Time handling functions are spread across multiple classes. Some
classes share the time handling functions.

This makes it hard to debug as it is hard to keep track of the multiple
time handling functions across the different files.

Extract the different time functions into a dedicated class.

Consolidating the location of time related functions makes it easy to
scale and debubg if there are any errors related to these functions.
