+++ src/main/java/duke/tasks/Task.java
@@ -1,11 +1,5 @@
 package duke.tasks;
 
-import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
-import java.time.format.DateTimeParseException;
-import java.util.Arrays;
-import java.util.List;
-
 /**
  * Basic template for tasks.
  */
@@ -39,33 +33,6 @@ public class Task {
      */
     public boolean isMatchingDescription(String match) {
         return this.description.contains(match);
-    }
-
-    /**
-     * Parses and converts string to LocalDateTime object.
-     *
-     * @param dateTime String containing date and time.
-     * @return dateTime     LocalDateTime object.
-     */
-    public LocalDateTime parseDateTime(String dateTime) {
-        List<String> separators = Arrays.asList("-", "/");
-        List<String> dateCombinations = Arrays.asList("dd_MM_yyyy ", "MM_dd_yyyy ", "yyyy_dd_MM ", "yyyy_MM_dd ");
-        List<String> timeCombinations = Arrays.asList("HH:mm", "HHmm", "hh:mm a");
-
-        for (String d : dateCombinations) {
-            for (String t : timeCombinations) {
-                for (String s : separators) {
-                    try {
-                        String pattern = d.replaceAll("_", s) + t;
-                        return LocalDateTime.parse(dateTime, DateTimeFormatter.ofPattern(pattern));
-                    } catch (DateTimeParseException dt) {
-                        // ...
-                    }
-                }
-            }
-        }
-
-        return null;
     }
 
     /**
